Rewrite `<ColumnBackButton>` and `<ColumnBackButtonSlim>` into Typescript
